rename pages directory
