+++ src/components/routes.jsx
@@ -1,8 +1,8 @@
 import { Route, Routes } from "react-router-dom";
-import { Home } from "../Pages/Home";
-import { Company } from "../Pages/Company";
+import { Home } from "../pages/Home";
+import { Company } from "../pages/Company";
 import { NotFoundPage } from "./NotFoundPage/NotFoundPage";
-import { MovieInfo } from "../Pages/MovieInfo";
+import { MovieInfo } from "../pages/MovieInfo";
 import { NewAccount } from "./CreateAccount/NewAccount";
 import { Login } from "./login/login";
